feat(bin-migration): Task 4 — prune/reorganize/remediation を scripts/lib/ に移設
- scripts/lib/prune.py: PLUGIN_ROOT に統一、lib.frontmatter → frontmatter
- scripts/lib/reorganize.py: PLUGIN_ROOT に統一、audit scripts path 削除
- scripts/lib/remediation.py: PLUGIN_ROOT に統一、インライン _plugin_root 全置換
- skills/*/scripts/*.py: importlib shim に変更（循環インポート回避）
